pyproject.toml
[project]
name = "flowbot3d"
version = "0.1.0"
description = "FlowBot3D: Learning 3D Articulation Flow to Manipulate Articulated Objects"
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE.txt"}
authors = [
  {email = "baeisner@andrew.cmu.edu", name = "Ben Eisner"},
  {email = "haolunz@andrew.cmu.edu", name = "Harry Zhang"},
  {email = "dheld@andrew.cmu.edu", name = "David Held"}
]
dependencies = [
  "gif",
  "gym <= 0.20.0",
  "h5py",
  "numpy <= 1.23.1",
  "open3d",
  "opencv-python",
  "pandas",
  "pybullet",
  "pyransac3d",
  "pytorch-lightning >= 1.6.4",
  "pytransform3d",
  "rpad-partnet-mobility-utils[pybullet] @ git+https://github.com/ipab-slmc/partnet_mobility_dataset",
  "rpad-pyg @ git+https://github.com/r-pad/pyg_libs",
  "rpad-visualize-3d @ git+https://github.com/r-pad/visualize_3d",
  "sapien == 1.1.0; sys_platform == 'linux'",
  "scipy",
  "shapely",
  "torch >= 1.11.0",
  "typer >= 0.4.0",
  "torch-scatter >= 2.0.9",
  "torch-sparse >= 0.6",
  "torch-cluster >= 1.6.0",
  "torch-spline-conv >= 1.2.1",
  "torch-geometric >= 2.0, < 2.3",
  "transforms3d",
  "trimesh",
  "tqdm",
  "wandb",
]

[build-system]
requires = [
  "setuptools >= 62.3.2",
  "setuptools-scm",
  "wheel",
]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
develop = [
  "autoflake",
  "black >= 22.3.0",
  "isort",
  "jupyterlab",
  "mypy",
  "pre-commit",
  "pylint",
  "pytest",
  "types-tqdm",
]
build_docs = [
  "mkdocs-material",
  "mkdocstrings[python]",
]

# This is required to allow us to have notebooks/ at the top level.
[tool.setuptools]
packages = ["flowbot3d"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.8
warn_return_any = true
warn_unused_configs = true

# Uncomment this when you have imports for mypy to ignore.
[[tool.mypy.overrides]]
module = [
  "cv2.*",
  "gif.*",
  "gym.*",
  "h5py.*",
  "mani_skill.*",
  "mani_skill_learn.*",
  "open3d.*",
  "pandas.*",
  "plotly.*",
  "pyransac3d.*",
  "sapien.*",
  "scipy.*",
  "torch_geometric.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--ignore=third_party/"
